Vitest batch 36: test-family-history (DOM-split) + test-emf-flow
test-family-history.js: source-inspection + getConditionsSummary
coverage ported to Vitest (56 asserts). The two DOM-runtime sections —
the apostrophe-condition round-trip probe (needs real innerHTML
parsing + dispatchEvent of an inline onmousedown) and the live
addFamilyHistoryEntry/deleteFamilyHistoryEntry handler test (reads
#fh-* inputs, re-renders renderDiagnosesModal against a live
#detail-modal) — split into test-family-history-dom.js on the
puppeteer runner.

test-emf-flow.js: full port (30 asserts). The CRUD/state assertions
are pure object mutations; the render/modal/photo/interpretation
paths are coverage-only (wrapped in try/catch or withTimeout,
asserting "X ran"), so they degrade gracefully against the Node DOM
stub. emf.js's render functions write into #detail-modal /
#modal-overlay — a local getElementById patch hands back a stub
element for those two IDs so renderEMFEditor's `modal.innerHTML`
doesn't throw on null.

Vitest: 83 → 85 tests. Puppeteer originals: 30 → 29.
