+++ tests/test-emf-flow.js
@@ -1,3 +1,4 @@
+#!/usr/bin/env node
 // test-emf-flow.js — Behavioral coverage of the EMF module (js/emf.js).
 //
 // Complements test-emf.js, which exercises the SBM-2015 threshold schema in
@@ -7,201 +8,225 @@
 // facade so V8 records every function as called, AND asserts the state
 // mutations a user would observe.
 //
-// Every async window function is wrapped in a hard 1.5s timeout — the
-// interpret/PDF flows open modals that wait for user input and would block
-// the runner indefinitely otherwise.
+// Run: node tests/test-emf-flow.js  (or via npm test)
+//
+// Full port — the CRUD/state assertions are pure object mutations; the
+// render / modal / photo / interpretation paths are coverage-only (wrapped
+// in try/catch or withTimeout, asserting only "X ran"), so they degrade
+// gracefully against the Node DOM stub. Every async window function is
+// wrapped in a hard 1.5s timeout — the interpret/PDF flows open modals that
+// wait for user input and would block the runner indefinitely otherwise.
 
-return (async () => {
-  let pass = 0, fail = 0;
-  const assert = (name, cond, detail) => {
-    if (cond) { pass++; console.log(`%c PASS %c ${name}`, 'background:#22c55e;color:#fff;padding:2px 6px;border-radius:3px', '', detail || ''); }
-    else { fail++; console.error(`%c FAIL %c ${name}`, 'background:#ef4444;color:#fff;padding:2px 6px;border-radius:3px', '', detail || ''); }
-  };
-  const withTimeout = (fn, ms = 1500) => Promise.race([
-    Promise.resolve().then(fn).catch(() => {}),
-    new Promise(r => setTimeout(r, ms)),
-  ]);
+import './_node-shim.js';
 
-  console.log('%c EMF Flow Tests ', 'background:#8b5cf6;color:#fff;font-size:14px;padding:4px 12px;border-radius:4px');
+// emf.js's render functions write into #detail-modal / #modal-overlay,
+// which exist in index.html's skeleton but not in the Node DOM stub. Hand
+// back a persistent stub element for those two IDs so renderEMFEditor's
+// `modal.innerHTML = html` doesn't throw on null — the render output is
+// coverage-only; every assertion here checks state mutations or "X ran".
+const _modalStub = () => ({
+  style: {}, classList: { add() {}, remove() {}, contains() { return false; }, toggle() {} },
+  appendChild() {}, remove() {}, setAttribute() {}, getAttribute() { return null; },
+  addEventListener() {}, removeEventListener() {},
+  querySelector() { return null; }, querySelectorAll() { return []; },
+  children: [], childNodes: [], innerHTML: '', textContent: '', value: '',
+});
+const _stubsById = {};
+const _origGetById = document.getElementById.bind(document);
+document.getElementById = (id) =>
+  (id === 'detail-modal' || id === 'modal-overlay')
+    ? (_stubsById[id] || (_stubsById[id] = _modalStub()))
+    : _origGetById(id);
 
-  // Bring in the actual modules — dynamic import forces parse + top-level
-  // execution, which is what registers the window facade.
-  const { state } = await import('/js/state.js');
-  await import('/js/emf.js?bust=' + Date.now());
-  await import('/js/data.js?bust=' + Date.now()); // saveImportedData lives here
+let pass = 0, fail = 0;
+const assert = (name, cond, detail) => {
+  if (cond) { pass++; console.log(`  PASS: ${name}`); }
+  else { fail++; console.log(`  FAIL: ${name}${detail ? ' — ' + detail : ''}`); }
+};
+const withTimeout = (fn, ms = 1500) => Promise.race([
+  Promise.resolve().then(fn).catch(() => {}),
+  new Promise(r => setTimeout(r, ms)),
+]);
 
-  assert('emf.js window facade loaded', typeof window.addEMFAssessment === 'function');
+console.log('=== EMF Flow Tests ===\n');
 
-  // Snapshot the existing emfAssessment subtree so we can restore it at the
-  // end and not pollute downstream tests (test-wearables-bp-merge has been
-  // observed to fail when state from this test bleeds over).
-  const _origEmf = state.importedData?.emfAssessment
-    ? JSON.parse(JSON.stringify(state.importedData.emfAssessment))
-    : null;
+// Bring in the actual modules — dynamic import forces parse + top-level
+// execution, which is what registers the window facade.
+const { state } = await import('../js/state.js');
+await import('../js/emf.js');
+await import('../js/data.js'); // saveImportedData lives here
 
-  state.importedData = state.importedData || {};
-  state.importedData.emfAssessment = { assessments: [], compareMode: false };
+assert('emf.js window facade loaded', typeof window.addEMFAssessment === 'function');
 
-  // ── 1. Assessment CRUD ────────────────────────────────────────────────
-  const beforeAdd = state.importedData.emfAssessment.assessments.length;
-  window.addEMFAssessment();
-  const afterAdd = state.importedData.emfAssessment.assessments.length;
-  assert('addEMFAssessment appends one assessment', afterAdd === beforeAdd + 1);
+// Snapshot the existing emfAssessment subtree so we can restore it at the
+// end and not pollute downstream tests (test-wearables-bp-merge has been
+// observed to fail when state from this test bleeds over).
+const _origEmf = state.importedData?.emfAssessment
+  ? JSON.parse(JSON.stringify(state.importedData.emfAssessment))
+  : null;
 
-  const asmId = state.importedData.emfAssessment.assessments[afterAdd - 1].id;
-  assert('New assessment has a string id', typeof asmId === 'string' && asmId.length > 0);
+state.importedData = state.importedData || {};
+state.importedData.emfAssessment = { assessments: [], compareMode: false };
 
-  window.updateEMFField(asmId, 'name', 'Coverage Probe');
-  const asm = state.importedData.emfAssessment.assessments.find(a => a.id === asmId);
-  assert('updateEMFField writes name', asm.name === 'Coverage Probe');
+// ── 1. Assessment CRUD ────────────────────────────────────────────────
+const beforeAdd = state.importedData.emfAssessment.assessments.length;
+window.addEMFAssessment();
+const afterAdd = state.importedData.emfAssessment.assessments.length;
+assert('addEMFAssessment appends one assessment', afterAdd === beforeAdd + 1);
 
-  window.updateEMFField(asmId, 'notes', 'Multi-line\nnotes here');
-  assert('updateEMFField writes notes', asm.notes === 'Multi-line\nnotes here');
+const asmId = state.importedData.emfAssessment.assessments[afterAdd - 1].id;
+assert('New assessment has a string id', typeof asmId === 'string' && asmId.length > 0);
 
-  // ── 2. Room CRUD ──────────────────────────────────────────────────────
-  // A new assessment ships with one default room (newRoom() inside emf.js).
-  const startingRooms = asm.rooms.length;
-  assert('New assessment has a default room', startingRooms >= 1);
+window.updateEMFField(asmId, 'name', 'Coverage Probe');
+const asm = state.importedData.emfAssessment.assessments.find(a => a.id === asmId);
+assert('updateEMFField writes name', asm.name === 'Coverage Probe');
 
-  window.addEMFRoom(asmId);
-  assert('addEMFRoom adds room', asm.rooms.length === startingRooms + 1);
+window.updateEMFField(asmId, 'notes', 'Multi-line\nnotes here');
+assert('updateEMFField writes notes', asm.notes === 'Multi-line\nnotes here');
 
-  window.updateEMFRoom(asmId, 0, 'name', 'Bedroom');
-  assert('updateEMFRoom updates name', asm.rooms[0].name === 'Bedroom');
+// ── 2. Room CRUD ──────────────────────────────────────────────────────
+// A new assessment ships with one default room (newRoom() inside emf.js).
+const startingRooms = asm.rooms.length;
+assert('New assessment has a default room', startingRooms >= 1);
 
-  window.updateEMFRoom(asmId, 0, 'location', 'east-facing wall');
-  assert('updateEMFRoom updates location', asm.rooms[0].location === 'east-facing wall');
+window.addEMFRoom(asmId);
+assert('addEMFRoom adds room', asm.rooms.length === startingRooms + 1);
 
-  // ── 3. Measurement + meter flow ──────────────────────────────────────
-  // updateEMFMeasurement stores `{ value, unit, meter }` — not the raw number.
-  // updateEMFMeter writes into the SAME nested object's `.meter` field, so
-  // the measurement must exist first.
-  window.updateEMFMeasurement(asmId, 0, 'acElectric', 12);
-  assert('updateEMFMeasurement stores nested value object',
-    asm.rooms[0].measurements?.acElectric?.value === 12);
-  assert('updateEMFMeasurement also tags the unit',
-    typeof asm.rooms[0].measurements.acElectric.unit === 'string');
+window.updateEMFRoom(asmId, 0, 'name', 'Bedroom');
+assert('updateEMFRoom updates name', asm.rooms[0].name === 'Bedroom');
 
-  window.updateEMFMeasurement(asmId, 0, 'rfMicrowave', 250);
-  window.updateEMFMeasurement(asmId, 0, 'acMagnetic', 80);
-  window.updateEMFMeasurement(asmId, 0, 'dirtyElectricity', 40);
-  assert('Multiple measurement types coexist',
-    Object.keys(asm.rooms[0].measurements).length >= 4);
+window.updateEMFRoom(asmId, 0, 'location', 'east-facing wall');
+assert('updateEMFRoom updates location', asm.rooms[0].location === 'east-facing wall');
 
-  window.updateEMFMeter(asmId, 0, 'acElectric', 'Safe and Sound EM3');
-  assert('updateEMFMeter writes into measurement.meter',
-    asm.rooms[0].measurements.acElectric.meter === 'Safe and Sound EM3');
+// ── 3. Measurement + meter flow ──────────────────────────────────────
+// updateEMFMeasurement stores `{ value, unit, meter }` — not the raw number.
+// updateEMFMeter writes into the SAME nested object's `.meter` field, so
+// the measurement must exist first.
+window.updateEMFMeasurement(asmId, 0, 'acElectric', 12);
+assert('updateEMFMeasurement stores nested value object',
+  asm.rooms[0].measurements?.acElectric?.value === 12);
+assert('updateEMFMeasurement also tags the unit',
+  typeof asm.rooms[0].measurements.acElectric.unit === 'string');
 
-  // Clear path: passing '' deletes the measurement.
-  window.updateEMFMeasurement(asmId, 0, 'dirtyElectricity', '');
-  assert('updateEMFMeasurement with empty value clears',
-    asm.rooms[0].measurements.dirtyElectricity === undefined);
+window.updateEMFMeasurement(asmId, 0, 'rfMicrowave', 250);
+window.updateEMFMeasurement(asmId, 0, 'acMagnetic', 80);
+window.updateEMFMeasurement(asmId, 0, 'dirtyElectricity', 40);
+assert('Multiple measurement types coexist',
+  Object.keys(asm.rooms[0].measurements).length >= 4);
 
-  // ── 4. Selection + render (need minimal modal DOM) ───────────────────
-  for (const id of ['modal-overlay', 'detail-modal']) {
-    if (!document.getElementById(id)) {
-      const el = document.createElement('div'); el.id = id; el.style.display = 'none'; document.body.appendChild(el);
-    }
+window.updateEMFMeter(asmId, 0, 'acElectric', 'Safe and Sound EM3');
+assert('updateEMFMeter writes into measurement.meter',
+  asm.rooms[0].measurements.acElectric.meter === 'Safe and Sound EM3');
+
+// Clear path: passing '' deletes the measurement.
+window.updateEMFMeasurement(asmId, 0, 'dirtyElectricity', '');
+assert('updateEMFMeasurement with empty value clears',
+  asm.rooms[0].measurements.dirtyElectricity === undefined);
+
+// ── 4. Selection + render (need minimal modal DOM) ───────────────────
+for (const id of ['modal-overlay', 'detail-modal']) {
+  if (!document.getElementById(id)) {
+    const el = document.createElement('div'); el.id = id; el.style.display = 'none'; document.body.appendChild(el);
   }
+}
 
-  try { window.openEMFAssessmentEditor(); } catch (_) {}
-  assert('openEMFAssessmentEditor ran', true);
+try { window.openEMFAssessmentEditor(); } catch (_) {}
+assert('openEMFAssessmentEditor ran', true);
 
-  try { window.toggleEMFAssessment(asmId); } catch (_) {}
-  assert('toggleEMFAssessment ran', true);
+try { window.toggleEMFAssessment(asmId); } catch (_) {}
+assert('toggleEMFAssessment ran', true);
 
-  try { window.selectEMFRoom(asmId, 0); } catch (_) {}
-  assert('selectEMFRoom ran', true);
+try { window.selectEMFRoom(asmId, 0); } catch (_) {}
+assert('selectEMFRoom ran', true);
 
-  await withTimeout(() => window.handleEMFRoomDropdown(asmId, 0, '0', { value: '0' }));
-  assert('handleEMFRoomDropdown ran', true);
+await withTimeout(() => window.handleEMFRoomDropdown(asmId, 0, '0', { value: '0' }));
+assert('handleEMFRoomDropdown ran', true);
 
-  // Compare view: needs ≥ 2 assessments. Add a second.
-  window.addEMFAssessment();
-  const secondId = state.importedData.emfAssessment.assessments.at(-1).id;
-  try { window.toggleEMFCompare(); } catch (_) {}
-  assert('toggleEMFCompare ran (with 2 assessments)', true);
-  try { window.toggleEMFCompare(); } catch (_) {} // toggle off
+// Compare view: needs ≥ 2 assessments. Add a second.
+window.addEMFAssessment();
+const secondId = state.importedData.emfAssessment.assessments.at(-1).id;
+try { window.toggleEMFCompare(); } catch (_) {}
+assert('toggleEMFCompare ran (with 2 assessments)', true);
+try { window.toggleEMFCompare(); } catch (_) {} // toggle off
 
-  // ── 5. Photos (FileReader path) ──────────────────────────────────────
-  // 1×1 PNG so the read actually succeeds (otherwise the photo never lands
-  // and removeEMFPhoto's index would be invalid).
-  const tinyPng = new Uint8Array([
-    0x89,0x50,0x4e,0x47,0x0d,0x0a,0x1a,0x0a, 0x00,0x00,0x00,0x0d,0x49,0x48,0x44,0x52,
-    0x00,0x00,0x00,0x01,0x00,0x00,0x00,0x01, 0x08,0x06,0x00,0x00,0x00,0x1f,0x15,0xc4,
-    0x89,0x00,0x00,0x00,0x0a,0x49,0x44,0x41, 0x54,0x78,0x9c,0x63,0x00,0x01,0x00,0x00,
-    0x05,0x00,0x01,0x0d,0x0a,0x2d,0xb4,0x00, 0x00,0x00,0x00,0x49,0x45,0x4e,0x44,0xae,
-    0x42,0x60,0x82,
-  ]);
-  const photoFile = new File([tinyPng], 'probe.png', { type: 'image/png' });
-  await withTimeout(() => window.addEMFPhotos(asmId, 0, [photoFile]));
-  assert('addEMFPhotos ran', true);
+// ── 5. Photos (FileReader path) ──────────────────────────────────────
+// 1×1 PNG so the read actually succeeds (otherwise the photo never lands
+// and removeEMFPhoto's index would be invalid).
+const tinyPng = new Uint8Array([
+  0x89,0x50,0x4e,0x47,0x0d,0x0a,0x1a,0x0a, 0x00,0x00,0x00,0x0d,0x49,0x48,0x44,0x52,
+  0x00,0x00,0x00,0x01,0x00,0x00,0x00,0x01, 0x08,0x06,0x00,0x00,0x00,0x1f,0x15,0xc4,
+  0x89,0x00,0x00,0x00,0x0a,0x49,0x44,0x41, 0x54,0x78,0x9c,0x63,0x00,0x01,0x00,0x00,
+  0x05,0x00,0x01,0x0d,0x0a,0x2d,0xb4,0x00, 0x00,0x00,0x00,0x49,0x45,0x4e,0x44,0xae,
+  0x42,0x60,0x82,
+]);
+const photoFile = new File([tinyPng], 'probe.png', { type: 'image/png' });
+await withTimeout(() => window.addEMFPhotos(asmId, 0, [photoFile]));
+assert('addEMFPhotos ran', true);
 
-  try { window.viewEMFPhoto(asmId, 0, 0); } catch (_) {}
-  assert('viewEMFPhoto ran', true);
+try { window.viewEMFPhoto(asmId, 0, 0); } catch (_) {}
+assert('viewEMFPhoto ran', true);
 
-  try { window.removeEMFPhoto(asmId, 0, 0); } catch (_) {}
-  assert('removeEMFPhoto ran', true);
+try { window.removeEMFPhoto(asmId, 0, 0); } catch (_) {}
+assert('removeEMFPhoto ran', true);
 
-  // ── 6. Interpretation flow (stub the AI; bound with a timeout) ───────
-  // streamInterpretation calls window.callClaudeAPI internally. interpret*
-  // functions open modals that wait on user clicks — they don't return
-  // promises, but their internal streamInterpretation IS async. Stub the AI
-  // so it resolves immediately; the modal stays open until we don't care
-  // anymore (closed by closeEMFInterpretation below).
-  const origCallAI = window.callClaudeAPI;
-  window.callClaudeAPI = async () => ({ text: 'Stub interpretation', usage: { inputTokens: 1, outputTokens: 1 } });
-  try { window.interpretEMFAssessment(asmId); } catch (_) {}
-  assert('interpretEMFAssessment ran', true);
-  try { window.interpretEMFComparison(); } catch (_) {}
-  assert('interpretEMFComparison ran', true);
-  // Drain microtasks so the stubbed AI promises resolve.
-  await new Promise(r => setTimeout(r, 50));
-  window.callClaudeAPI = origCallAI;
+// ── 6. Interpretation flow (stub the AI; bound with a timeout) ───────
+// streamInterpretation calls window.callClaudeAPI internally. interpret*
+// functions open modals that wait on user clicks — they don't return
+// promises, but their internal streamInterpretation IS async. Stub the AI
+// so it resolves immediately; the modal stays open until we don't care
+// anymore (closed by closeEMFInterpretation below).
+const origCallAI = window.callClaudeAPI;
+window.callClaudeAPI = async () => ({ text: 'Stub interpretation', usage: { inputTokens: 1, outputTokens: 1 } });
+try { window.interpretEMFAssessment(asmId); } catch (_) {}
+assert('interpretEMFAssessment ran', true);
+try { window.interpretEMFComparison(); } catch (_) {}
+assert('interpretEMFComparison ran', true);
+// Drain microtasks so the stubbed AI promises resolve.
+await new Promise(r => setTimeout(r, 50));
+window.callClaudeAPI = origCallAI;
 
-  try { window.closeEMFInterpretation(); } catch (_) {}
-  assert('closeEMFInterpretation ran', true);
+try { window.closeEMFInterpretation(); } catch (_) {}
+assert('closeEMFInterpretation ran', true);
 
-  try { window.discussEMFInterpretation(); } catch (_) {}
-  assert('discussEMFInterpretation ran', true);
+try { window.discussEMFInterpretation(); } catch (_) {}
+assert('discussEMFInterpretation ran', true);
 
-  // ── 7. PDF import path (stubbed) ─────────────────────────────────────
-  const origParsePDF = window.parsePDFFile;
-  window.parsePDFFile = async () => 'EMF assessment\nBedroom\nacElectric: 12 V/m\n';
-  window.callClaudeAPI = async () => ({ text: JSON.stringify({ assessments: [] }), usage: { inputTokens: 1, outputTokens: 1 } });
-  const fakePdf = new File([new Uint8Array(10)], 'probe.pdf', { type: 'application/pdf' });
-  await withTimeout(() => window.handleEMFPDF(fakePdf));
-  assert('handleEMFPDF ran', true);
-  window.parsePDFFile = origParsePDF;
-  window.callClaudeAPI = origCallAI;
+// ── 7. PDF import path (stubbed) ─────────────────────────────────────
+const origParsePDF = window.parsePDFFile;
+window.parsePDFFile = async () => 'EMF assessment\nBedroom\nacElectric: 12 V/m\n';
+window.callClaudeAPI = async () => ({ text: JSON.stringify({ assessments: [] }), usage: { inputTokens: 1, outputTokens: 1 } });
+const fakePdf = new File([new Uint8Array(10)], 'probe.pdf', { type: 'application/pdf' });
+await withTimeout(() => window.handleEMFPDF(fakePdf));
+assert('handleEMFPDF ran', true);
+window.parsePDFFile = origParsePDF;
+window.callClaudeAPI = origCallAI;
 
-  // ── 8. removeEMFRoom + deleteEMFAssessment ───────────────────────────
-  window.addEMFRoom(asmId);
-  const beforeRm = asm.rooms.length;
-  try { window.removeEMFRoom(asmId, asm.rooms.length - 1); } catch (_) {}
-  assert('removeEMFRoom decrements room count', asm.rooms.length === beforeRm - 1);
+// ── 8. removeEMFRoom + deleteEMFAssessment ───────────────────────────
+window.addEMFRoom(asmId);
+const beforeRm = asm.rooms.length;
+try { window.removeEMFRoom(asmId, asm.rooms.length - 1); } catch (_) {}
+assert('removeEMFRoom decrements room count', asm.rooms.length === beforeRm - 1);
 
-  // deleteEMFAssessment awaits showConfirmDialog (imported directly from
-  // utils.js — ES module bindings are read-only, so we can't stub it
-  // post-import). The fn opens a real overlay dialog that nobody clicks;
-  // the await hangs until withTimeout cancels. The function IS entered
-  // (V8 marks it called), which is the coverage goal. We just need an
-  // assertion that doesn't depend on the actual delete happening.
-  await withTimeout(() => window.deleteEMFAssessment(secondId));
-  await withTimeout(() => window.deleteEMFAssessment(asmId));
-  assert('deleteEMFAssessment called without throwing', true);
+// deleteEMFAssessment awaits showConfirmDialog (imported directly from
+// utils.js — ES module bindings are read-only, so we can't stub it
+// post-import). The fn opens a real overlay dialog that nobody clicks;
+// the await hangs until withTimeout cancels. The function IS entered
+// (V8 marks it called), which is the coverage goal. We just need an
+// assertion that doesn't depend on the actual delete happening.
+await withTimeout(() => window.deleteEMFAssessment(secondId));
+await withTimeout(() => window.deleteEMFAssessment(asmId));
+assert('deleteEMFAssessment called without throwing', true);
 
-  // ── 9. saveEMFExplicit ───────────────────────────────────────────────
-  try { window.saveEMFExplicit(); } catch (_) {}
-  assert('saveEMFExplicit ran', true);
+// ── 9. saveEMFExplicit ───────────────────────────────────────────────
+try { window.saveEMFExplicit(); } catch (_) {}
+assert('saveEMFExplicit ran', true);
 
-  // Restore the snapshot so downstream tests see the same emfAssessment
-  // they expected. This is the load-bearing cleanup — without it,
-  // test-wearables-bp-merge has been observed to fail because saveEMFExplicit
-  // persisted our probe data over its expected fixtures.
-  if (_origEmf) state.importedData.emfAssessment = _origEmf;
-  else delete state.importedData.emfAssessment;
+// Restore the snapshot so downstream tests see the same emfAssessment
+// they expected. This is the load-bearing cleanup — without it,
+// test-wearables-bp-merge has been observed to fail because saveEMFExplicit
+// persisted our probe data over its expected fixtures.
+if (_origEmf) state.importedData.emfAssessment = _origEmf;
+else delete state.importedData.emfAssessment;
 
-  console.log(`\n%c EMF Flow Result: ${pass} passed, ${fail} failed `,
-    `background:${fail ? '#ef4444' : '#22c55e'};color:#fff;font-size:13px;padding:3px 10px;border-radius:3px`);
-})();
+console.log(`\nResults: ${pass} passed, ${fail} failed, ${pass + fail} total`);
+process.exit(fail > 0 ? 1 : 0);
